+++ HISTORY.md
@@ -2,6 +2,6 @@
 
 ## v1.0.0
 
-### 変更
-- [konpeito](https://github.com/natade-jp/konpeito) ver6.0.1 をベースに `WSH JScript` 用にライブラリを分離
-- 統計計算部分（`Complex`, `Matrix` 含む）のみ抽出しました
+### 新規作成
+- [konpeito](https://github.com/natade-jp/konpeito) ver6.0.1 をベースに `WSH JScript` でも動作するように `ES3` 向けに抽出
+- 抽出対象は、統計計算部分（`Complex`, `Matrix` 含む）のみ